requirements.txt
ultralytics==8.3.144
ultralytics-thop==2.0.14
opencv-python==4.11.0.86
requests==2.32.3
python-dotenv==1.1.0
google-generativeai==0.8.5
